+++ src/main/java/com/ifmo/lessons/diploma/service/ProductService.java
@@ -1,6 +1,8 @@
 package com.ifmo.lessons.diploma.service;
 
+import com.ifmo.lessons.diploma.entity.CustomerProduct;
 import com.ifmo.lessons.diploma.entity.Product;
+import com.ifmo.lessons.diploma.repository.CustomerProductRepository;
 import com.ifmo.lessons.diploma.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,10 +18,12 @@ import java.util.stream.Collectors;
 @Service
 public class ProductService {
     private final ProductRepository repository;
+    private final CustomerProductRepository customerProductRepository;
 
     @Autowired
-    public ProductService(ProductRepository repository) {
+    public ProductService(ProductRepository repository, CustomerProductRepository customerProductRepository) {
         this.repository = repository;
+        this.customerProductRepository = customerProductRepository;
     }
 
     public List<Product> getAll() {
@@ -41,7 +45,11 @@ public class ProductService {
     }
 
     public void delete(Product product) {
-        // ToDo: используется ли товар у клиента
+        // Проверить используется ли товар у клиента
+        List<CustomerProduct> list = customerProductRepository.findAllByProduct(product.getId());
+        if (!list.isEmpty()) {
+            throw new IllegalArgumentException("Товар используется у клиента " + list.get(0).getCustomer().getName() + ". Удаление невозможно.");
+        }
         repository.delete(product);
     }
 }
